refactor: Separate tastingSheets get method into a custom hook
